@@ -1,9 +1,11 @@
-const Environment = require('./Environment');
-const Organism = require('../Organism/Organism');
-const GridMap = require('../Grid/GridMap');
-const Renderer = require('../Rendering/Renderer');
-const EditorController = require("../Controllers/EditorController");
-const Species = require('../Stats/Species');
+import Environment from './Environment';
+import Organism from '../Organism/Organism';
+import GridMap from '../Grid/GridMap';
+import Renderer from '../Rendering/Renderer';
+import EditorController from "../Controllers/EditorController";
+import Species from '../Stats/Species';
+import { Mouth } from '../Organism/Cell/BodyCells/BodyCells';
+import { Empty } from '../Organism/Cell/EnvironmentCells/EnvironmentCells';
 
 class OrganismEditor extends Environment{
     constructor(registry) {
@@ -13,7 +15,7 @@ class OrganismEditor extends Environment{
         this.renderer = new Renderer('editor-canvas', 'editor-env', cell_size);
         this.controller = new EditorController(this, this.renderer.canvas);
         this.Registry.Cells.Subscribe('register', (payload) => console.log(payload))
-        this.grid_map = new GridMap(this, this.Registry.Cells, 15, 15, cell_size);
+        this.grid_map = new GridMap(this, 15, 15, cell_size);
         this.clear();
     }
 
@@ -23,8 +25,8 @@ class OrganismEditor extends Environment{
         }
     }
 
-    changeCell(c, r, state, owner) {
-        super.changeCell(c, r, state, owner);
+    changeCell(c, r, type, owner) {
+        super.changeCell(c, r, type, owner);
         this.renderFull();
     }
 
@@ -33,17 +35,17 @@ class OrganismEditor extends Environment{
     }
 
     // absolute c r, not local
-    addCellToOrg(c, r, state) {
+    addCellToOrg(c, r, type) {
         var center = this.grid_map.getCenter();
         var loc_c = c - center[0];
         var loc_r = r - center[1];
         var prev_cell = this.organism.anatomy.getLocalCell(loc_c, loc_r)
         if (prev_cell != null) {
-            var new_cell = this.organism.anatomy.replaceCell(state, prev_cell.loc_col, prev_cell.loc_row, false);
-            this.changeCell(c, r, state, new_cell);
+            var new_cell = this.organism.anatomy.replaceCell(type, prev_cell.loc_col, prev_cell.loc_row, false);
+            this.changeCell(c, r, type, new_cell);
         }
         else if (this.organism.anatomy.canAddCellAt(loc_c, loc_r)){
-            this.changeCell(c, r, state, this.organism.anatomy.addDefaultCell(state, loc_c, loc_r));
+            this.changeCell(c, r, type, this.organism.anatomy.addDefaultCell(type, loc_c, loc_r));
         }
         this.organism.species = new Species(this.organism.anatomy, null, 0);
     }
@@ -59,14 +61,14 @@ class OrganismEditor extends Environment{
         var prev_cell = this.organism.anatomy.getLocalCell(loc_c, loc_r)
         if (prev_cell != null) {
             if (this.organism.anatomy.removeCell(loc_c, loc_r)) {
-                this.changeCell(c, r, this.Registry.Cells.GetByName('empty'), null);
+                this.changeCell(c, r, Empty, null);
                 this.organism.species = new Species(this.organism.anatomy, null, 0);
             }
         }
     }
 
     setOrganismToCopyOf(orig_org) {
-        this.grid_map.fillGrid(this.Registry.Cells.GetByName('empty'));
+        this.grid_map.fillGrid(Empty);
         var center = this.grid_map.getCenter();
         this.organism = new Organism(center[0], center[1], this, orig_org);
         this.organism.updateGrid();
@@ -80,13 +82,13 @@ class OrganismEditor extends Environment{
     }
 
     clear() {
-        this.grid_map.fillGrid(this.Registry.Cells.GetByName('empty'));
+        this.grid_map.fillGrid(Empty);
         var center = this.grid_map.getCenter();
         this.organism = new Organism(center[0], center[1], this, null);
-        this.organism.anatomy.addDefaultCell(this.Registry.Cells.GetByName('mouth'), 0, 0);
+        this.organism.anatomy.addDefaultCell(Mouth, 0, 0);
         this.organism.updateGrid();
         this.organism.species = new Species(this.organism.anatomy, null, 0);
     }
 }
 
-module.exports = OrganismEditor;
+export default OrganismEditor;
