@@ -1,7 +1,11 @@
-const BodyCell = require("./BodyCell");
-const Hyperparams = require("../../../Hyperparameters");
+import BodyCell from "./BodyCell";
+import HyperParameters from "../../../Hyperparameters";
+import { BodyCellState } from "../CellState";
+import { Armor, Killer } from "./BodyCells";
 
 class KillerCell extends BodyCell{
+    static state = new BodyCellState(KillerCell, "killer", "", "");
+
     constructor(org, loc_col, loc_row){
         super(org, loc_col, loc_row);
     }
@@ -10,7 +14,7 @@ class KillerCell extends BodyCell{
         var env = this.org.env;
         var c = this.getRealCol();
         var r = this.getRealRow();
-        for (var loc of Hyperparams.killableNeighbors) {
+        for (var loc of HyperParameters.killableNeighbors) {
             var cell = env.grid_map.cellAt(c+loc[0], r+loc[1]);
             this.killNeighbor(cell);
         }
@@ -18,14 +22,14 @@ class KillerCell extends BodyCell{
 
     killNeighbor(n_cell) {
         // console.log(n_cell)
-        if(n_cell == null || n_cell.owner == null || n_cell.owner == this.org || !n_cell.owner.living || n_cell.state == this.org.env.Registry.GetState('armor')) 
+        if(n_cell == null || n_cell.owner == null || n_cell.owner == this.org || !n_cell.owner.living || n_cell.type == Armor) 
             return;
-        var is_hit = n_cell.state == this.org.env.Registry.GetState('killer'); // has to be calculated before death
+        var is_hit = n_cell.type == Killer; // has to be calculated before death
         n_cell.owner.harm();
-        if (Hyperparams.instaKill && is_hit) {
+        if (HyperParameters.instaKill && is_hit) {
             this.org.harm();
         }
     }
 }
 
-module.exports = KillerCell;
+export default KillerCell;
