@@ -1,30 +1,33 @@
-const BodyCell = require("./BodyCell");
-const Hyperparams = require("../../../Hyperparameters");
+import BodyCell from "./BodyCell";
+import HyperParameters from "../../../Hyperparameters";
+import { BodyCellState } from "../CellState";
+import { Empty, Food } from "../EnvironmentCells/EnvironmentCells";
 
 class ProducerCell extends BodyCell{
+    static state = new BodyCellState(ProducerCell, "producer", "", "");
     constructor(org, loc_col, loc_row){
         super(org, loc_col, loc_row);
         this.org.anatomy.is_producer = true;
     }
 
     performFunction() {
-        if (this.org.anatomy.is_mover && !Hyperparams.moversCanProduce)
+        if (this.org.anatomy.is_mover && !HyperParameters.moversCanProduce)
             return;
         var env = this.org.env;
-        var prob = Hyperparams.foodProdProb;
+        var prob = HyperParameters.foodProdProb;
         var real_c = this.getRealCol();
         var real_r = this.getRealRow();
         if (Math.random() * 100 <= prob) {
-            var loc = Hyperparams.growableNeighbors[Math.floor(Math.random() * Hyperparams.growableNeighbors.length)]
+            var loc = HyperParameters.growableNeighbors[Math.floor(Math.random() * HyperParameters.growableNeighbors.length)]
             var loc_c=loc[0];
             var loc_r=loc[1];
-            var cell = env.grid_map.cellAt(real_c+loc_c, real_r+loc_r);
-            if (cell != null && cell.state == env.Registry.Cells.GetByName('empty')){
-                env.changeCell(real_c+loc_c, real_r+loc_r, env.Registry.Cells.GetByName('food'), null);
+            var gridCell = env.grid_map.cellAt(real_c+loc_c, real_r+loc_r);
+            if (gridCell != null && gridCell.type == Empty){
+                env.changeCell(real_c+loc_c, real_r+loc_r, Food, null);
                 return;
             }
         }
     }
 }
 
-module.exports = ProducerCell;
+export default ProducerCell;
