@@ -1,34 +1,35 @@
-const Directions = {
-    up:0,
-    right:1,
-    down:2,
-    left:3,
-    scalars:[[0,-1],[1,0],[0,1],[-1,0]],
-    getRandomDirection: function() {
-        return Math.floor(Math.random() * 4);
-    },
-    getRandomScalar: function() {
-        return this.scalars[Math.floor(Math.random() * this.scalars.length)];
-    },
-    getOppositeDirection: function(dir) {
-        switch(dir){
-            case this.up:
-                return this.down;
-            case this.down:
-                return this.up;
-            case this.left:
-                return this.right;
-            case this.right:
-                return this.left;
-        }
-    },
-    rotateRight: function(dir) {
-        dir++;
-        if (dir > 3){
-            dir = 0;
-        }
-        return dir;
+
+export const up = 0;
+export const right = 1;
+export const down = 2;
+export const left = 3;
+export const scalars = [[0, -1], [1, 0], [0, 1], [-1, 0]];
+
+export function getRandomDirection() {
+    return Math.floor(Math.random() * 4);
+}
+
+export function getRandomScalar() {
+    return scalars[Math.floor(Math.random() * scalars.length)];
+}
+
+export function getOppositeDirection(dir) {
+    switch (dir) {
+        case up:
+            return down;
+        case down:
+            return up;
+        case left:
+            return right;
+        case right:
+            return left;
     }
 }
 
-module.exports = Directions;
+export function rotateRight(dir) {
+    dir++;
+    if (dir > 3) {
+        dir = 0;
+    }
+    return dir;
+}
