@@ -1,11 +1,3 @@
-const { CellState, BodyCellState } = require('./Organism/Cell/CellState');
-const ArmorCell = require("./Organism/Cell/BodyCells/ArmorCell");
-const EyeCell = require("./Organism/Cell/BodyCells/EyeCell");
-const KillerCell = require("./Organism/Cell/BodyCells/KillerCell");
-const MouthCell = require("./Organism/Cell/BodyCells/MouthCell");
-const MoverCell = require("./Organism/Cell/BodyCells/MoverCell");
-const ProducerCell = require("./Organism/Cell/BodyCells/ProducerCell");
-
 class Registry {
     constructor() {
         this.subscribers = [];
@@ -47,60 +39,20 @@ class CellRegistry extends Registry {
         this.Register(item);
     }
 
-    GetByType(type) {
-        return this.registry.find(x => x.CellType && x.CellType === type);
-    }
-
     GetByName(name) {
-        return this.registry.find(x => x.name === name);
+        return this.registry.find(x => x.state.name === name);
     }
 
     WithTag(tag) {
-        return this.registry.filter(x => x.tags.find(t => t === tag));
+        return this.registry.filter(x => x.state.tags.find(t => t === tag));
     }
 }
-
-
 
 class LifeEngineRegistries {
-    constructor() {
-        this.states = [];
-        this.Cells = new CellRegistry();
-        
-        
-    }
-
-    getRandomName() {
-        var names = this.Cells.All().map(x => x.name);
-        return names[Math.floor(Math.random() * names.length)];
-    }
-
-    getRandomLivingType() {
-        const living = this.Cells.WithTag('living');
-        return living[Math.floor(Math.random() * living.length)];
-    }
-
-    States() {
-        return this.Cells.All();
-    }
-
-    LivingStates() {
-        return this.Cells.WithTag('living');
-    }
-
-    GetState(name) {
-        return this.Cells.GetByName(name);
-    }
-
-    RegisterEnvironmentCell(name, color) {
-        var state = new CellState(name);
-        this.Cells.RegisterCell(new CellState(name, color, "", []));
-    }
-
-    RegisterBodyCell(name, color, cellType) {
-        var state = new CellState(name, 'living');
-        this.Cells.RegisterCell(new BodyCellState(cellType, name, color, "", ['living']));
+    constructor(cells) {
+        this.Cells = cells;
     }
 }
 
-module.exports = LifeEngineRegistries;
+export const Cells = new CellRegistry();
+export const Registries = new LifeEngineRegistries(Cells);
