@@ -1,4 +1,4 @@
-const FossilRecord = require("../FossilRecord");
+import FossilRecord from "../FossilRecord";
 
 class ChartController {
     constructor(title, y_axis="", note="") {
@@ -81,4 +81,4 @@ class ChartController {
     }
 }
 
-module.exports = ChartController;
+export default ChartController;
