@@ -16,11 +16,8 @@
 //          ...
 //          ...
 
-const Neighbors = {
-    all: [[0, 1],[0, -1],[1, 0],[-1, 0],[-1, -1],[1, 1],[-1, 1],[1, -1]],
-    adjacent: [[0, 1],[0, -1],[1, 0],[-1, 0]],
-    corners: [[-1, -1],[1, 1],[-1, 1],[1, -1]],
-    allSelf: [[0, 0],[0, 1],[0, -1],[1, 0],[-1, 0],[-1, -1],[1, 1],[-1, 1],[1, -1]]
-}
+export const all = [[0, 1],[0, -1],[1, 0],[-1, 0],[-1, -1],[1, 1],[-1, 1],[1, -1]];
+export const adjacent = [[0, 1],[0, -1],[1, 0],[-1, 0]];
+export const corners = [[-1, -1],[1, 1],[-1, 1],[1, -1]];
+export const allSelf = [[0, 0],[0, 1],[0, -1],[1, 0],[-1, 0],[-1, -1],[1, 1],[-1, 1],[1, -1]];
 
-module.exports = Neighbors;
