@@ -1,12 +1,25 @@
+const webpack = require('webpack');
 const path = require('path');
 
 module.exports = {
+    devtool: 'source-map',
     entry: './src/index.js',
     output: {
         filename: 'bundle.js',
         path: path.resolve(__dirname, 'dist/js/'),
     },
-    externals: {
-        jquery: 'jQuery'
-    }
+    module: {
+        rules: [
+          {
+            test: /\.(js)$/,
+            use: "babel-loader",
+          },
+        ],
+      },
+      plugins: [
+        new webpack.ProvidePlugin({
+            $: "jquery",
+            jQuery: "jquery"
+        })
+    ]
 };
